Integrated table into the main application. Now accesses token from localStorage where it was set by the login page.
